+++ test/e2e/specs/contact.js
@@ -8,7 +8,6 @@ module.exports = {
       .assert.containsText('@contactForm', 'Email is required.')
       .assert.containsText('@contactForm', 'Phone is required.')
       .assert.containsText('@contactForm', 'Description is required.')
-      .assert.urlEquals(browser.page.contact().url())
       .setValue('@email', 'notanaccount')
       .assert.containsText('@contactForm', 'Please enter a valid email address.')
       .api.end()