+++ test/unit/specs/ContactForm.spec.js
@@ -1,0 +1,13 @@
+import Vue from 'vue'
+import ContactForm from 'src/components/ContactForm'
+
+describe('ContactForm.vue', () => {
+  it('should render field labels', () => {
+    const vm = new Vue({
+      el: document.createElement('div'),
+      render: (h) => h(ContactForm)
+    })
+    expect(vm.$el.querySelector('label[for=name]').textContent)
+      .to.contain('Your name')
+  })
+})